Isolate packet methods
