+++ crates/relayer-cosmos/src/methods/mod.rs
@@ -2,4 +2,5 @@ pub mod channel;
 pub mod connection;
 pub mod consensus_state;
 pub mod create_client;
+pub mod packet;
 pub mod update_client;
